Initial Commit
I'm attempting to convert my console version of this program into one
that uses a Windows Forms graphical interface.  With a LOT of great help
from a friend of mine, I've been able to get this interface as far as I
have thus far.  It is up to me to finish fixing all of the other holes
in the program to make it fully functional.